split client & server into transport agnostic packages
`client-core` and `server-core` will be able to be run over any transport implementing the `Socket` interface from `client-server-common`

`client-websocket` and `server-websocket` provide websocket implementations of `Socket`
