+++ js/examples/todomvc/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@types/ws": "^8.5.3",
     "@vlcn.io/rx-tbl": "workspace:*",
-    "@vlcn.io/sync-client": "workspace:*",
+    "@vlcn.io/client-websocket": "workspace:*",
     "@vlcn.io/wa-crsqlite": "workspace:*",
     "nanoid": "^4.0.0",
     "react": "^18.2.0",
