+++ js/sync-reference/client-core/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@vlcn.io/client-core",
+  "type": "module",
+  "version": "0.5.3",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "module": "dist/index.js",
+  "main": "dist/index.js",
+  "devDependencies": {
+    "typescript": "^4.9.3"
+  },
+  "scripts": {
+    "build": "tsc --build",
+    "watch": "tsc --build --watch",
+    "deep-clean": "rm -rf ./dist || true && rm tsconfig.tsbuildinfo || true",
+    "start": "pnpm build && node dist/main.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:vlcn-io/cr-sqlite.git",
+    "directory": "https://github.com/vlcn-io/cr-sqlite/tree/main/js/sync-client"
+  },
+  "dependencies": {
+    "@vlcn.io/client-server-common": "workspace:*",
+    "@vlcn.io/rx-tbl": "workspace:*",
+    "@vlcn.io/xplat-api": "workspace:*"
+  }
+}
