+++ js/sync-reference/client-core/tsconfig.json
