+++ js/sync-reference/client-websocket/src/logger.ts
