+++ js/sync-reference/client-websocket/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../tsconfig-template.json",
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "rootDir": "./src"
+  },
+  "include": ["./src/"],
+  "references": [
+    { "path": "../client-server-common" },
+    { "path": "../../xplat-api" },
+    { "path": "../../rx-tbl" }
+  ]
+}
