+++ js/sync-reference/server-core/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@vlcn.io/server-core",
+  "type": "module",
+  "version": "0.5.6",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "module": "dist/index.js",
+  "main": "dist/index.js",
+  "devDependencies": {
+    "@types/better-sqlite3": "^7.6.2",
+    "@types/node": "^18.11.9",
+    "@types/uuid": "^8.3.4",
+    "typescript": "^4.9.3",
+    "vitest": "^0.25.2"
+  },
+  "scripts": {
+    "build": "tsc --build",
+    "watch": "tsc --build --watch",
+    "test": "vitest run",
+    "deep-clean": "rm -rf ./dist || true && rm tsconfig.tsbuildinfo || true",
+    "start": "pnpm build && node dist/bin/main.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:vlcn-io/cr-sqlite.git",
+    "directory": "https://github.com/vlcn-io/cr-sqlite/tree/main/js/sync-reference/server"
+  },
+  "dependencies": {
+    "@vlcn.io/client-server-common": "workspace:*",
+    "@vlcn.io/crsqlite": "workspace:*",
+    "better-sqlite3": "^8.0.1",
+    "import-meta-resolve": "^2.1.0",
+    "nanoid": "^4.0.0",
+    "uuid": "^9.0.0",
+    "winston": "^3.8.2",
+    "yargs": "^17.6.2"
+  }
+}
