+++ js/sync-reference/server-websocket/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@vlcn.io/sync-server",
+  "name": "@vlcn.io/server-websocket",
   "type": "module",
   "version": "0.5.6",
   "files": [
@@ -34,7 +34,7 @@
   },
   "dependencies": {
     "@types/yargs": "^17.0.18",
-    "@vlcn.io/client-server-common": "workspace:*",
+    "@vlcn.io/server-core": "workspace:*",
     "@vlcn.io/crsqlite": "workspace:*",
     "better-sqlite3": "^8.0.1",
     "cors": "^2.8.5",
