+++ js/sync-reference/server-websocket/src/bin/main.ts
@@ -7,10 +7,8 @@ import express from "express";
 import { IncomingMessage } from "node:http";
 import { WebSocketServer } from "ws";
 import * as http from "http";
-import { Connection } from "../connection.js";
-import logger from "../logger.js";
+import { Connection, logger, contextStore } from "@vlcn.io/server-core";
 import { nanoid } from "nanoid";
-import contextStore from "../contextStore.js";
 import yargs from "yargs/yargs";
 import { hideBin } from "yargs/helpers";
 import { configure } from "../config.js";
