+++ js/sync-reference/server-websocket/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../tsconfig-template.json",
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "rootDir": "./src",
+    "allowSyntheticDefaultImports": true
+  },
+  "include": ["./src/"],
+  "references": [
+    { "path": "../../node-allinone" },
+    { "path": "../server-core" }
+  ]
+}
