+++ pnpm-workspace.yaml
@@ -4,8 +4,10 @@ packages:
   - "js/browser/wa-crsqlite"
   - "deps/wa-sqlite"
   - "js/sync-reference/p2p"
-  - "js/sync-reference/server"
-  - "js/sync-reference/client"
+  - "js/sync-reference/server-core"
+  - "js/sync-reference/server-websocket"
+  - "js/sync-reference/client-core"
+  - "js/sync-reference/client-websocket"
   - "js/sync-reference/client-server-common"
   - "js/tests/xplat-tests"
   - "js/tests/nodeno-tests"
